fix tests, move most errors to shared code
