--- common/script/api-v3/customError.js
@@ -1,8 +1,0 @@
-// Base class for custom application errors
-// It extends Error and capture the stack trace
-export default class CustomError extends Error {
-  constructor () {
-    super();
-    Error.captureStackTrace(this, this.constructor);
-  }
-}
